Overhaul of sbxjoin and sbxsplit

+++ sbxjoin.sh
@@ -3,8 +3,8 @@
 shopt -s extglob
 
 if (( $# < 2 )); then
-  echo "Usage : in_prefix out_file"
-  exit 1
+    echo "Usage : in_prefix out_file"
+    exit 1
 fi
 
 in_prefix=$1
@@ -18,58 +18,62 @@ rm -rf $out_container
 
 echo "Decoding parts"
 
-for file in $in_prefix.part+([0-9]).sbx; do
-  echo "  Decoding $file"
+for file in $in_prefix.part*; do
+    if [[ $file == *.sbx ]]; then
+        echo "  Decoding $file"
 
-  output=$(rsbx sort --json $file $file.sorted --burst 0)
-  error=$(echo $output | jq -r ".error")
-  if [[ $error != null ]]; then
-    echo "    Error occured during sorting"
-    echo "    $error"
-    exit 1
-  fi
+        output=$(rsbx sort --json $file $file.sorted --burst 0)
+        error=$(echo $output | jq -r ".error")
+        if [[ $error != null ]]; then
+            echo "    Error occured during sorting"
+            echo "    $error"
+            exit 1
+        fi
 
-  output=$(rsbx repair --json $file.sorted)
-  error=$(echo $output | jq -r ".error")
-  if [[ $error != null ]]; then
-    echo "    Error occured during repairing"
-    echo "    $error"
-    exit 1
-  fi
+        output=$(rsbx repair --json $file.sorted)
+        error=$(echo $output | jq -r ".error")
+        if [[ $error != null ]]; then
+            echo "    Error occured during repairing"
+            echo "    $error"
+            exit 1
+        fi
 
-  output=$(rsbx decode --json $file.sorted)
-  if [[ $error != "null" ]]; then
-    echo "    Error occured during encoding"
-    echo "    $error"
-    exit 1
-  fi
+        output=$(rsbx decode --json $file.sorted)
+        if [[ $error != "null" ]]; then
+            echo "    Error occured during encoding"
+            echo "    $error"
+            exit 1
+        fi
 
-  recorded_hash=$(echo $output | jq -r ".stats.recordedHash")
-  output_hash=$(echo $output | jq -r ".stats.hashOfOutputFile")
+        recorded_hash=$(echo $output | jq -r ".stats.recordedHash")
+        output_hash=$(echo $output | jq -r ".stats.hashOfOutputFile")
 
-  if [[ $recorded_hash != $output_hash ]]; then
-    echo "    Error : hash mismatch for $file"
-    exit 1
-  fi
+        if [[ $recorded_hash != $output_hash ]]; then
+            echo "    Error : hash mismatch for $file"
+            exit 1
+        fi
 
-  rm $file.sorted
+        rm $file.sorted
+    fi
 done
 
 echo "Concatenating parts"
 for file in $in_prefix.part+([0-9]); do
-  echo "  Adding $file to final container"
+    echo "  Adding $file to final container"
 
-  cat $file >> $out_container
-  rm $file
+    cat $file >> $out_container
+    if [ -f $file.sbx ]; then
+        rm $file
+    fi
 done
 
 echo "Sorting container"
 output=$(rsbx sort --json $out_container $out_container.sorted --burst 0)
 error=$(echo $output | jq -r ".error")
 if [[ $error != null ]]; then
-  echo "  Error occured during sorting"
-  echo "  $error"
-  exit 1
+    echo "  Error occured during sorting"
+    echo "  $error"
+    exit 1
 fi
 
 mv $out_container.sorted $out_container
@@ -78,23 +82,23 @@ echo "Repairing container"
 output=$(rsbx repair --json $out_container)
 error=$(echo $output | jq -r ".error")
 if [[ $error != null ]]; then
-  echo "  Error occured during repairing"
-  echo "  $error"
-  exit 1
+    echo "  Error occured during repairing"
+    echo "  $error"
+    exit 1
 fi
 repairs_failed=$(echo $output | jq -r ".stats.numberOfBlocksFailedToRepairData")
 if (( $repairs_failed > 0 )); then
-  echo "  Failed to repair container"
-  echo "  The decoded data will have missing data"
+    echo "  Failed to repair container"
+    echo "  The decoded data will have missing data"
 fi
 
 echo "Decoding container"
 output=$(rsbx decode --json $out_container $out_file)
 error=$(echo $output | jq -r ".error")
 if [[ $error != null ]]; then
-  echo "  Error occured during decoding"
-  echo "  $error"
-  exit 1
+    echo "  Error occured during decoding"
+    echo "  $error"
+    exit 1
 fi
 
 recorded_hash=$(echo $output | jq -r ".stats.recordedHash")
@@ -102,10 +106,10 @@ output_hash=$(echo $output | jq -r ".stats.hashOfOutputFile")
 
 echo "Checking hash"
 if [[ $recorded_hash != $output_hash ]]; then
-  echo "  Error : Output file hash mismatch"
-  echo "  Recorded    hash : $recorded_hash"
-  echo "  Output file hash : $output_hash"
-  exit 1
+    echo "  Error : Output file hash mismatch"
+    echo "  Recorded    hash : $recorded_hash"
+    echo "  Output file hash : $output_hash"
+    exit 1
 fi
 
 # clean up
